fix(audio): keep the Volume x87 chain in f64 and label the VERA-only guards
`VocClass::ReadINI @ 0x00750440` reads `Volume=` through
`CCINIClass::ReadDouble @ 0x005283D0`, which returns a **double** (the
sscanf'd float is widened at `0x0052855D`/`0x00528569`, and the `%` form
multiplies by the double 0.01 at `0x007E3808` with no float store). The
product then never leaves the x87 stack between `0x00750507 FMUL float
0.01f` and `0x0075053F CALL Math::ftol` -- there is no intermediate `FSTP
float` in that range, and no game code narrows the precision control, so
the truncation sees at least 53 significant bits.

VERA narrowed the product to f32 first. Because `0.01f` is just below
0.01, the exact product sits just under the round value and the f32
rounding pushes it up, so `Volume=` 25/50/75/100 and the `Volume=5000%`
form each landed one linear step high (4096/8192/12288/16384 instead of
4095/8191/12287/16383) -- 55 stock soundmd.ini sections, 0.14-0.59 dB too
loud. Worse, `volume_and_min_volume_follow_read_double_and_the_clamps`
pinned 16384 and 8192 as if they were the decompile golden; both are
re-pinned to the native values, with a new test covering all four
quarter-volumes.

`read_double` now returns f64, `volume_linear` takes f64 and keeps the
whole chain there, and `min_volume_fraction` reproduces the one narrowing
the native MinVolume path really has (`0x007505A8 FST float [ESP+8]`,
with the high clamp still testing the un-narrowed product).

Also labels or removes the VERA-only guards inside gamemd-derived bodies:

- `SoundEntry::range` widens u16 -> i32; `AudioEventClass::SetRange @
  0x004065E0` is `MOV [ECX+0x50], EDX` with no clamp, so the u16 clamp is
  deleted rather than labelled.
- An entry whose `Sounds=` yields no usable token is now registered with
  zero samples instead of skipped: native's `VocClass` exists and plays
  nothing, where VERA's skip let the audio-bag fallback play a file gamemd
  never would. The remaining `Sounds=`-as-section-filter is labelled
  VERA-internal (native walks `[SoundList]`, VERA has no such pass).
- `combine_linear`, `native_volume_amplitude` and `pan_channel_gains`
  clamps labelled VERA-internal: `FUN_004010C0`'s volume path is
  `SHR/SHR/IMUL/SHR` with no bound (only its *pan* path clamps), and
  `FUN_0040A6D0` indexes its table unchecked. Unreachable from any VERA
  producer; kept because Rust would panic on the fixed 101-entry table.
- `PlayShifts::shifted_sample_rate`'s `.max(1)` labelled VERA-internal:
  `FUN_00401190` is a plain truncating divide; the floor only guards
  VERA's own resampler and needs `FShift=` below -100 to fire.

Verified this session by decompile/disassemble: 0x005283D0, 0x00750440,
0x007504D0..0x00750600, 0x00750AC0, 0x007510D0, 0x004065C0..0x00406660,
0x004010C0, 0x00401190, and read_memory on 0x007E1718, 0x007E1748,
0x007E2AC8, 0x007E3808, 0x007E5168, 0x007EAAE0, 0x007EF38C.

+++ src/audio/sfx.rs
@@ -177,7 +177,7 @@ pub struct SpatialSource {
 impl SpatialSource {
     pub fn from_entry(entry: &SoundEntry) -> Self {
         Self {
-            range_cells: i32::from(entry.range),
+            range_cells: entry.range,
             type_flags: entry.type_flags,
             min_volume: entry.min_volume,
         }
@@ -275,9 +275,10 @@ pub fn calc_volume_and_pan(
     let full_w: f32 = half_w + half_w;
     // `LEA EAX,[EAX+EAX*2]; LEA EAX,[EAX+EAX*4]; SHL EAX,2` at
     // `0x00750B0F..0x00750B17` — a 32-bit ×60 that wraps rather than trapping.
-    // `SoundEntry::range` is a `u16`, so the product cannot exceed 3_932_100;
-    // `wrapping_mul` is here to keep a hand-built `SpatialSource` on the
-    // native truncation instead of a debug-build panic.
+    // `AudioEventClass::SetRange @ 0x004065E0` stores `Range=` as a full int,
+    // so `wrapping_mul` is the transcription of the native overflow, not a
+    // guard: it is what keeps a `Range=` past 35_791_394 on the native
+    // truncation instead of panicking in a debug build.
     let max_range: f32 = (source.range_cells.wrapping_mul(RANGE_MULTIPLIER)) as f32;
 
     if source.type_flags & sound_type::SHROUD != 0 && shrouded {
@@ -370,12 +371,34 @@ fn attenuation_amplitude(hundredths_db: i16) -> f32 {
 
 /// Product of two native linear volumes: `DSoundBuffer::CombineInterps
 /// FUN_004010C0` (`0x004010C7..0x004010D6`), `(a * b) >> 14`.
+///
+/// **VERA-internal, gamemd equivalent UNCHECKED: both `.clamp`s.** The native
+/// volume path is `SHR EDX,0x10; SHR EAX,0x10; IMUL EDX,EAX; SHR EDX,0xe`
+/// (`disassemble_function 0x004010C0`, read this session) — no bound at all;
+/// the two operands are merely the top halves of 32-bit fixed-point fields, so
+/// native accepts anything up to `0xFFFF` on each side and can return well
+/// past `0x4000`. (The *pan* path at `0x00401116..0x0040114C` does clamp to
+/// `0..0x4000`; the volume path does not, so this is not borrowed from there.)
+/// Trigger: an operand outside `0..=0x4000`, which no VERA producer can make —
+/// `SoundEntry::volume_linear` comes out of the native `[0, 1]` clamp times
+/// 16384, and [`PlayShifts::volume_linear`] is `0x4000 - ((vshift << 14)/100)`
+/// with `vshift` already held to `0..=100` by native `SetVShift @ 0x00406620`.
+/// Player effect: none reachable. Frequency: never. Downstream risk: the guard
+/// is what keeps [`native_volume_amplitude`]'s fixed 101-entry table index in
+/// bounds, where Rust would panic and gamemd would read past the table.
 pub fn combine_linear(a: i32, b: i32) -> i32 {
     (a.clamp(0, VOLUME_SCALE) * b.clamp(0, VOLUME_SCALE)) >> 14
 }
 
 /// Amplitude the DirectSound layer produces for one combined linear volume:
 /// `FUN_0040A6D0` indexes the table with `volume * 25 >> 12` (0..=100).
+///
+/// VERA-internal, gamemd equivalent UNCHECKED: the `.clamp`. `FUN_0040A6D0`
+/// computes `(v >> 16) * 25 >> 12` and indexes `DAT_00816380` with it
+/// unchecked, so an out-of-range volume reads past the 101-entry table in
+/// gamemd; in Rust that is a panic, which is not a behaviour worth
+/// reproducing. Trigger and frequency: as [`combine_linear`] — unreachable
+/// from any VERA producer.
 pub fn native_volume_amplitude(linear: i32) -> f32 {
     let index = (linear.clamp(0, VOLUME_SCALE) * 25) >> 12;
     attenuation_amplitude(DSOUND_ATTENUATION_TABLE[index as usize])
@@ -387,6 +410,12 @@ pub fn native_volume_amplitude(linear: i32) -> f32 {
 /// `p = (pan * 25 >> 11) - 100` and calls `SetPan(table[100 - |p|])` for a
 /// left pan (negative DirectSound pan attenuates the right channel) and
 /// `SetPan(-table[100 - p])` for a right pan (attenuates the left channel).
+///
+/// VERA-internal, gamemd equivalent UNCHECKED: the `.clamp`, for the same
+/// reason as [`native_volume_amplitude`] — native indexes the table
+/// unchecked. `CalcVolumeAndPan` produces the pan as
+/// `ftol(clamp(offsetX, -fullW, fullW) * 8192 / fullW + 8192)`, which is
+/// already `0..=0x4000`, so nothing in VERA can reach the guard.
 pub fn pan_channel_gains(pan: i32) -> (f32, f32) {
     let p = ((pan.clamp(0, PAN_SCALE) * 25) >> 11) - 100;
     let attenuated = attenuation_amplitude(DSOUND_ATTENUATION_TABLE[(100 - p.abs()) as usize]);
@@ -511,7 +540,18 @@ impl PlayShifts {
         }
     }
 
-    /// Playback rate: `FUN_00401190` returns `(pct * rate) / 100`.
+    /// Playback rate: `FUN_00401190` returns `(pct * rate) / 100`
+    /// (`SHR ECX,0x10; IMUL ECX,EDX;` then the `0x51EB851F` magic divide by
+    /// 100, truncating toward zero).
+    ///
+    /// VERA-internal, gamemd equivalent UNCHECKED: the `.max(1)`. The native
+    /// body has no floor — it hands the raw quotient to the DirectSound
+    /// buffer's `SetFrequency`, which rejects 0 at the device layer. VERA
+    /// instead feeds the rate to its own resampler, where 0 is not a
+    /// well-defined input. Trigger: `frequency_pct <= 0`, i.e. a `FShift=`
+    /// whose low bound is at or below -100; the widest stock `FShift=` is
+    /// `-15 15`. Player effect: none reachable. Frequency: never on retail
+    /// data. Downstream risk: none.
     pub fn shifted_sample_rate(&self, sample_rate: u32) -> u32 {
         ((i64::from(self.frequency_pct) * i64::from(sample_rate)) / 100).max(1) as u32
     }
